feed_selector and compute draws logic added

+++ massa-node/src/main.rs
@@ -319,11 +319,13 @@ async fn launch(
         .await
         .expect("could not start network controller");
 
-    // give the controller to final state in order for it to feed the cycles
-    final_state
-        .write()
-        .compute_initial_draws()
-        .expect("could not compute initial draws"); // TODO: this might just mean a bad bootstrap, no need to panic, just reboot
+    if args.restart_from_snapshot_at_period.is_none() {
+        // give the controller to final state in order for it to feed the cycles
+        final_state
+            .write()
+            .compute_initial_draws()
+            .expect("could not compute initial draws"); // TODO: this might just mean a bad bootstrap, no need to panic, just reboot
+    }
 
     // Storage costs constants
     let storage_costs_constants = StorageCostsConstants {
